+++ naga/src/back/hlsl/mod.rs
@@ -119,7 +119,7 @@ use core::fmt::Error as FmtError;
 
 use thiserror::Error;
 
-use crate::{back, proc};
+use crate::{back, ir, proc};
 
 #[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize))]
@@ -442,9 +442,12 @@ pub struct ReflectionInfo {
 pub struct PipelineOptions {
     /// The entry point to write.
     ///
+    /// Entry points are identified by a shader stage specification,
+    /// and a name.
+    ///
     /// If `None`, all entry points will be written. If `Some` and the entry
     /// point is not found, an error will be thrown while writing.
-    pub entry_point: Option<String>,
+    pub entry_point: Option<(ir::ShaderStage, String)>,
 }
 
 #[derive(Error, Debug)]
@@ -461,8 +464,8 @@ pub enum Error {
     Override,
     #[error(transparent)]
     ResolveArraySizeError(#[from] proc::ResolveArraySizeError),
-    #[error("entry point with name '{0}' not found")]
-    EntryPointNotFound(String),
+    #[error("entry point with stage {0:?} and name '{1}' not found")]
+    EntryPointNotFound(ir::ShaderStage, String),
 }
 
 #[derive(PartialEq, Eq, Hash)]